+++ CHANGELOG.md
@@ -16,6 +16,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
   - E.g. `slumber show paths config` prints just the config path
 - Add `--edit` flag to `slumber show config` and `slumber show collection`
   - This will open the global config/collection file in your configured editor, similar to `git config --edit`
+- `slumber import` now accepts a URL in addition to a file path
+  - If a URL is given, the file will be downloaded and imported
 
 ## Fixed